+++ tests/git_edit_index_tests.py
@@ -269,8 +269,14 @@ class ReflectIndexChangeTests(unittest.TestCase, WithPatching):
         self.perform_git_action = mock.Mock()
         self.patch('git_edit_index.perform_git_action', self.perform_git_action)
 
+        self.os_path_isdir = mock.Mock()
+        self.patch('git_edit_index.os.path.isdir', self.os_path_isdir)
+
         self.os_remove = mock.Mock()
         self.patch('git_edit_index.os.remove', self.os_remove)
+
+        self.shutil_rmtree = mock.Mock()
+        self.patch('git_edit_index.shutil.rmtree', self.shutil_rmtree)
 
     def test_performs_correct_action_when_untracked_file_is_to_be_added(self):
         orig_entry = IndexEntry('?', 'file.txt')
@@ -283,10 +289,21 @@ class ReflectIndexChangeTests(unittest.TestCase, WithPatching):
     def test_performs_correct_action_when_untracked_file_is_to_be_deleted(self):
         orig_entry = IndexEntry('?', 'file.txt')
         new_entry = NoIndexEntry('file.txt')
+        self.os_path_isdir.return_value = False
 
         reflect_index_change(orig_entry, new_entry)
 
         self.os_remove.assert_called_once_with('file.txt')
+        self.assertFalse(self.perform_git_action.called)
+
+    def test_performs_correct_action_when_untracked_directory_is_to_be_deleted(self):
+        orig_entry = IndexEntry('?', 'dir')
+        new_entry = NoIndexEntry('dir')
+        self.os_path_isdir.return_value = True
+
+        reflect_index_change(orig_entry, new_entry)
+
+        self.shutil_rmtree.assert_called_once_with('dir')
         self.assertFalse(self.perform_git_action.called)
 
     def test_performs_correct_action_when_modified_file_is_to_be_added(self):